+++ composer/src/lib.rs
@@ -212,6 +212,24 @@ impl<'de> Deserialize<'de> for ComposerRepositories {
                         })
                         .map_err(|_| {
                             Error::custom(format!("invalid value at object key '{key}', expected repository definition object or disablement using boolean `false`"))
+                        })
+                        .and_then(|mut repository| {
+                            match repository {
+                                ComposerRepository::Composer { ref mut name, .. }
+                                | ComposerRepository::Path { ref mut name, .. }
+                                | ComposerRepository::Package { ref mut name, .. }
+                                | ComposerRepository::Url { ref mut name, .. }
+                                | ComposerRepository::Other { ref mut name, .. }
+                                => {
+                                    if name.is_some() {
+                                        Err(Error::custom(format!("invalid value at object key '{key}', repository definition must not have a 'name' field")))
+                                    } else {
+                                        name.replace(key);
+                                        Ok(repository)
+                                    }
+                                },
+                                ComposerRepository::Disabled(_) => Ok(repository)
+                            }
                         })?;
                     values.push(repo);
                 }
@@ -241,6 +259,8 @@ impl ComposerRepositories {
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct ComposerBasePackage {
+    #[serde(rename = "_comment")]
+    pub comment: Option<Value>,
     #[serde(rename = "type")]
     pub kind: Option<String>,
     pub abandoned: Option<ComposerPackageAbandoned>,
@@ -471,6 +491,7 @@ pub struct ComposerPackageFunding {
 pub enum ComposerRepository {
     #[serde(rename_all = "kebab-case")]
     Composer {
+        name: Option<String>,
         #[serde(rename = "type")]
         kind: MustBe!("composer"),
         url: ComposerUrlOrPathUrl, // can also be a relative path
@@ -484,6 +505,7 @@ pub enum ComposerRepository {
     },
     #[serde(rename_all = "kebab-case")]
     Path {
+        name: Option<String>,
         #[serde(rename = "type")]
         kind: MustBe!("path"),
         url: PathBuf,
@@ -494,6 +516,7 @@ pub enum ComposerRepository {
     },
     #[serde(rename_all = "kebab-case")]
     Package {
+        name: Option<String>,
         #[serde(rename = "type")]
         kind: MustBe!("package"),
         #[serde_as(as = "OneOrMany<_, PreferOne>")]
@@ -505,6 +528,7 @@ pub enum ComposerRepository {
     // any other repo type that has a URL field
     #[serde(rename_all = "kebab-case")]
     Url {
+        name: Option<String>,
         #[serde(rename = "type")]
         kind: String,
         url: ComposerUrlOrScpLikeOrPathUrl, // can be an SCP style SSH "URL" or relative path, too
@@ -516,6 +540,7 @@ pub enum ComposerRepository {
     },
     #[serde(rename_all = "kebab-case")]
     Other {
+        name: Option<String>,
         #[serde(rename = "type")]
         kind: String,
         canonical: Option<bool>,
@@ -537,6 +562,7 @@ impl ComposerRepository {
         options: Option<Map<String, Value>>,
     ) -> Self {
         Self::Path {
+            name: None,
             #[allow(clippy::default_trait_access)]
             kind: Default::default(),
             url: path.into(),
@@ -549,6 +575,7 @@ impl ComposerRepository {
 impl From<Vec<ComposerPackage>> for ComposerRepository {
     fn from(value: Vec<ComposerPackage>) -> Self {
         Self::Package {
+            name: None,
             #[allow(clippy::default_trait_access)]
             kind: Default::default(),
             package: value,
@@ -847,7 +874,47 @@ mod tests {
     }
 
     #[rstest]
-    fn test_composer_json(#[files("tests/fixtures/*.json")] path: PathBuf) {
+    fn test_composer_json(
+        #[files("tests/fixtures/*.json")]
+        #[exclude(r"\.xfail\.json$")]
+        path: PathBuf,
+    ) {
+        let composer_json = std::io::BufReader::new(fs::File::open(&path).unwrap());
+        let package: ComposerRootPackage = serde_json::from_reader(composer_json).unwrap();
+        // The test cases may have special testing related instructions in the _comment field
+        if let Some(Value::Array(ref comments)) = package.package.comment {
+            comments
+                .iter()
+                .filter_map(|comment| match comment {
+                    Value::Object(comment) => Some(comment),
+                    _ => None,
+                })
+                .flat_map(std::iter::IntoIterator::into_iter)
+                .for_each(|comment| match comment {
+                    (op, Value::String(details)) if op == "compare-deserialized" => {
+                        // We are supposed to compare the deserialized form to that of another file
+                        // This is useful for testing e.g. normalizations, such as for the repositories notations
+                        let mut their_path = path.clone();
+                        their_path.set_file_name(details);
+                        let mut ours = package.clone();
+                        let mut theirs: ComposerRootPackage = serde_json::from_reader(
+                            std::io::BufReader::new(fs::File::open(their_path).unwrap()),
+                        )
+                        .unwrap();
+                        // The _comment fields obviously differ, so we discard them before comparing
+                        ours.package.comment = None;
+                        theirs.package.comment = None;
+                        assert_json_diff::assert_json_eq!(ours, theirs);
+                    }
+                    _ => panic!("unexpected magic testing comment instruction"),
+                });
+        }
+    }
+
+    #[allow(clippy::should_panic_without_expect)]
+    #[should_panic]
+    #[rstest]
+    fn test_broken_composer_json(#[files("tests/fixtures/*.xfail.json")] path: PathBuf) {
         let composer_json = fs::read(&path).unwrap();
         serde_json::from_slice::<ComposerRootPackage>(&composer_json).unwrap();
     }