+++ buildpacks/php/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Filter output from Composer on platform installation failure for better readability (for parity with the Classic buildpack)
+- Explicitly handle newly introduced `name` field in `composer.json` (version 2.9+) `repositories` definitions
 
 ## [1.1.0] - 2025-10-27